Log added in Project Apis

+++ src/main/java/com/naib/taskmanager/service/ProjectService.java
@@ -1,5 +1,6 @@
 package com.naib.taskmanager.service;
 
+import com.naib.taskmanager.controller.ProjectController;
 import com.naib.taskmanager.dto.ProjectDataDTO;
 import com.naib.taskmanager.dto.ProjectResponseDataDTO;
 import com.naib.taskmanager.model.Role;
@@ -8,19 +9,18 @@ import com.naib.taskmanager.model.dao.Task;
 import com.naib.taskmanager.model.dao.User;
 import com.naib.taskmanager.repository.ProjectRepository;
 import com.naib.taskmanager.repository.TaskRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 @Service
 public class ProjectService {
-
+    public static Logger LOG = LoggerFactory.getLogger(ProjectService.class);
     @Autowired
     UserService userService;
     @Autowired
@@ -84,12 +84,32 @@ public class ProjectService {
     }
 
     @Transactional
-    public void deleteByProjectId(Integer projectId) {
-        List<Task> tasks = taskRepository.findAllByProject_Id(projectId);
-        tasks.forEach(task -> {
-            task.setProject(null);
-        });
-        taskRepository.saveAll(tasks);
-        projectRepository.deleteById(projectId);
+    public ProjectResponseDataDTO deleteByProjectId(Integer projectId, HttpServletRequest request) {
+        boolean isUserPermittedToDeleteProject = false;
+        User loggedInUser = null;
+                ProjectResponseDataDTO responseDataDTO = new ProjectResponseDataDTO();
+        if (request.isUserInRole(Role.ADMIN.name())) {
+            loggedInUser = userService.loggedInUser(request);
+            isUserPermittedToDeleteProject = true;
+        } else if (request.isUserInRole(Role.USER.name())) {
+            loggedInUser = userService.loggedInUser(request);
+            Project project = projectRepository.findByIdAndUser(projectId,loggedInUser);
+            if (project != null){
+                isUserPermittedToDeleteProject = true;
+            }
+        }
+        if (isUserPermittedToDeleteProject){
+            List<Task> tasks = taskRepository.findAllByProject_Id(projectId);
+            tasks.forEach(task -> {
+                task.setProject(null);
+            });
+            taskRepository.saveAll(tasks);
+            projectRepository.deleteById(projectId);
+            LOG.info("Project with project_id: {} is deleted by user: {}", projectId, loggedInUser.getUsername());
+            responseDataDTO.setMessage("Project Deleted Successfully");
+        }else {
+            responseDataDTO.setMessage("Project Deletion Unsuccessful");
+        }
+       return  responseDataDTO;
     }
 }
